src/visualizer.py
"""
Cube visualization module.

Provides functions to generate visual representations of Rubik's Cube states.
"""

from typing import Optional, Tuple
import matplotlib.pyplot as plt
from PIL import Image
import numpy as np


class CubeVisualizer:
    """Handles visualization of Rubik's Cube states."""
    
    def __init__(self):
        """Initialize the visualizer with default settings."""
        pass
    
    def create_net_visualization(self, cube_state, output_path: str) -> None:
        """
        Create a 2D net visualization of the cube.
        
        Args:
            cube_state: The cube state to visualize
            output_path: Path to save the output image
        """
        pass
    
    def create_numbered_visualization(self, cube_state, output_path: str) -> None:
        """
        Create a visualization with numbered positions (1-54).
        
        Args:
            cube_state: The cube state to visualize
            output_path: Path to save the output image
        """
        pass
    
    def export_as_png(self, figure, output_path: str) -> None:
        """Export visualization as PNG."""
        pass
    
    def export_as_svg(self, figure, output_path: str) -> None:
        """Export visualization as SVG."""
        pass