requirements.txt
# Core dependencies
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Optional: Advanced solving
# kociemba>=1.2.1  # Uncomment if implementing Kociemba algorithm